+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonForm.cs
@@ -1577,16 +1577,28 @@ public class KryptonForm : VisualForm,
     {
         base.OnHandleCreated(e);
 
-        if (IsMdiContainer && !_mdiTransferred)
+        // Differ on MdiContainer first
+        if (IsMdiContainer)
         {
-            SetInheritedControlOverride();
-
-            Control.ControlCollection checkForRibbon = _internalKryptonPanel.Controls;
-
-            for (var i = checkForRibbon.Count - 1; i >= 0; i--)
+            if (!_mdiTransferred)
             {
-                base.Controls.Add(checkForRibbon[i]);
+                SetInheritedControlOverride();
+
+                Control.ControlCollection checkForRibbon = _internalKryptonPanel.Controls;
+
+                for (var i = checkForRibbon.Count - 1; i >= 0; i--)
+                {
+                    base.Controls.Add(checkForRibbon[i]);
+                }
             }
+        }
+        else if (_internalPanelState == InheritBool.Inherit)
+        {
+            // #2448 | Work-around / fix | Only runs on non mdi containers
+            // This happens when KForm is instantiated manually without designer source,
+            // which runs the layout methods through InitializeComponent.
+            SuspendLayout();
+            ResumeLayout(false);
         }
 
         // Register with the ActiveFormTracker